requirements.txt
pytest == 8.1.1
selene == 2.0.0rc9
allure-pytest == 2.13.5